requirements.txt
flask-sqlalchemy
flask-migrate
pymysql
flask-jwt-extended
flask-bcrypt
python-dotenv
authlib